Print balances in test
