+++ basics/transfer-sol/poseidon/transfer-sol/package.json
@@ -6,9 +6,7 @@
     "ts-mocha": "ts-mocha"
   },
   "dependencies": {
-    "@coral-xyz/anchor": "^0.30.1",
-    "anchor-bankrun": "^0.5.0",
-    "solana-bankrun": "^0.4.0"
+    "@coral-xyz/anchor": "^0.30.1"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
