+++ basics/transfer-sol/poseidon/transfer-sol/tests/transfer-sol.ts
@@ -1,6 +1,7 @@
 import * as anchor from "@coral-xyz/anchor";
 import { Program } from "@coral-xyz/anchor";
 import { TransferSol } from "../target/types/transfer_sol";
+import { Keypair, LAMPORTS_PER_SOL, PublicKey } from "@solana/web3.js";
 
 describe("transfer-sol", () => {
   const provider = anchor.AnchorProvider.env();
@@ -9,15 +10,36 @@ describe("transfer-sol", () => {
   const program = anchor.workspace.TransferSol as Program<TransferSol>;
 
   it("Is SOL transferred!", async () => {
-    // Add your test here.
+    const payer = provider.wallet;
+    const recipient = anchor.web3.Keypair.generate();
+    const transferAmount = new anchor.BN(1000000);
+
+    await printBalance(payer.publicKey, recipient.publicKey, "Before Transfer");
+
     const tx = await program.methods
-      .transferSol(new anchor.BN(1000000))
+      .transferSol(transferAmount)
       .accounts({
-        payer: provider.wallet.publicKey,
-        recipient: anchor.web3.Keypair.generate().publicKey,
+        payer: payer.publicKey,
+        recipient: recipient.publicKey,
       })
       .rpc();
 
+    await printBalance(payer.publicKey, recipient.publicKey, "After Transfer");
+
     console.log("Your transaction signature", tx);
   });
+
+  async function printBalance(
+    payer: PublicKey,
+    recipient: PublicKey,
+    when: string
+  ) {
+    const payerBalance = await provider.connection.getBalance(payer);
+    const recipientBalance = await provider.connection.getBalance(recipient);
+    console.log(
+      `${when} - Payer: ${payerBalance / LAMPORTS_PER_SOL} - Recipient: ${
+        recipientBalance / LAMPORTS_PER_SOL
+      }`
+    );
+  }
 });
